requirements.txt
numpy
pandas
matplotlib
tensorflow
keras
opencv-python
python-dotenv
twilio
simpleaudio